% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_xlsx_to_cube.R
\name{crea_d}
\alias{crea_d}
\title{From excel_sheet to a cube}
\usage{
crea_d(
  name_flr = "excel",
  sheet.x = "INDEX_m",
  var.x = "IPC",
  language.x = c("English", "PosTxt_E")
)
}
\arguments{
\item{name_flr}{folder's name, by default: "excel"}

\item{sheet.x}{sheet's name, by default: "INDEX_m"}

\item{var.x}{column's name, by default: "var.x"}

\item{language.x}{set the language, by default: c("English", "PosTxt_E"). Other possibilities: "German",
"PosTxt_D"; "French", "PosTxt_F"; and "Italian", "PosTxt_I"}
}
\description{
Transforms "INDEX_m" sheet of file: "su-e-05.02.67.xlsx" in a long form dataset.
}
\details{
The function: \code{crea_d} is the second step of the swissPCI package. By default, it produces a dataset
with the data contained in the sheet: "INDEX_m". Of course, via the same FUN, it is possible to change
the sheet, among these possibilities: "INDEX_m", "VAR_m-1", "VAR_m-12", "CONTR_m".
}
\examples{
# Data were stored in a folder called: "my_pci"

\dontrun{
# Standard dataset, INDEX_m sheet and labels in english
d1 <- crea_d("my_pci")

# Dataset with: yearly variations ("VAR_m-12" sheet), in german
d2 <- crea_d(m_pci, "VAR_m-12", "VAR_y", c("German", "PosTxt_D"))

}

}
