% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_ALL_pci_to_cube.R
\name{get_all_PCI}
\alias{get_all_PCI}
\title{Get the full original swiss PCI dataset}
\usage{
get_all_PCI(
  sheet.x = c("VAR_m-1", "VAR_m-12", "CONTR_m"),
  var.x = c("VAR_m-1", "VAR_m-12", "CONTR_m"),
  name_flr = c("excel", "rda"),
  language.x = c("English", "PosTxt_E")
)
}
\arguments{
\item{sheet.x}{sheet's names: "VAR_m-1", "VAR_m-12", "CONTR_m"}

\item{var.x}{variable's names, by default like sheet's names}

\item{name_flr}{destination folder's name, by default: c("excel", "rda")}

\item{language.x, }{set the language, by default: c("English", "PosTxt_E"). Other possibilities: "German",
"PosTxt_D"; "French", "PosTxt_F"; and "Italian", "PosTxt_I"}
}
\description{
The function: \code{get_all_PCI} set up the original big cube, in a long form.
}
\details{
The function: \code{get_all_PCI} downloads and puts in a long form all the data
contained in the excel file "su-e-05.02.67.xlsx".
}
