% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_pci100_to_cube.R
\name{get_PCI}
\alias{get_PCI}
\title{Get the swiss PCI dataset in a .rds file (only index-base 100)}
\usage{
get_PCI(name_flr = c("excel", "rda"), language.x = c("English", "PosTxt_E"))
}
\arguments{
\item{name_flr}{names of the folders, one origin and one like destination, by default: c("excel", "rda")}

\item{language.x, }{set the language, by default: c("English", "PosTxt_E"). Other possibilities: "German",
"PosTxt_D"; "French", "PosTxt_F"; and "Italian", "PosTxt_I"}
}
\description{
The function: \code{get_PCI} set up the big cube with only the variable: "PCI", in a long form.
}
\details{
The function: \code{get_PCI} downloads and puts in a long form the data contained in the "sheet: "INDEX_m" sheet
of the excel file: "su-e-05.02.67.xlsx".
}
\examples{
# Data were stored in a folder called: "my_pci", the .rds file finish in a folder named: "my_rda"

\dontrun{
# Save a .rds file, in italian
get_PCI(c("my_pci", "my_rda"), c("Italian", "PosTxt_I"))

}

}
