% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_pci_var_contr.R
\name{PCI_plus}
\alias{PCI_plus}
\title{Create the variables: VAR and CONTR}
\usage{
PCI_plus(d, time.x = "month", k = 12, labels.x = c("PCI", "VAR_y", "CONTR_y"))
}
\arguments{
\item{d}{base PCI dataset}

\item{time.x}{period: "month","quarter" or "year", by default = "month",}

\item{k}{the lag: 1 or 12, or 4, by default = 12}

\item{labels.x}{label names, by default = c("PCI", "VAR_y, "CONTR_y")}
}
\description{
The function: \code{PCI_plus} add the variables: "VAR" and "CONTR".
}
\details{
The function: \code{PCI_plus} generate the variables: "VAR" and "CONTR", calculated from the variables: "value" (index_100)
multiplied with "weigth" (column: "w2022").
}
\examples{

\dontrun{
# The PCI dataset "swissPCI.rds" is stored in my folder: "my_rds"
d <- readRDS("my_rds/swissPCI.rds")

# Create a dataset with: "quarter" data and with change on previous quarter (lag = 1)
d2 <- PCI_plus(d, "quarter", 1, c("PCI", "VAR_q", "CONTR_q"))

}

}
