% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03b_pci_level.R
\name{PCI_plus0}
\alias{PCI_plus0}
\title{Level column harmonization}
\usage{
PCI_plus0(d, labels.x = c("Level_0", "Level", "wCat"))
}
\arguments{
\item{d}{base PCI dataset}

\item{labels.x}{labels colnames, by default = c("Level_0", "Level", "wCat")}
}
\description{
The function: \code{PCI_plus0} add two new columns: "Level", "wCat
}
\details{
The function: \code{PCI_plus0l_wCat} add two new columns: "Level", "wCat. Starting from the original column: "Level"
(renamed: "Level_0") and from the specific "PosNo" of energy products and those of seasonal products.
}
\examples{

\dontrun{
d0 <- readRDS("my_rds/swissPCI.rds")
d1 <- PCI__plus0(d0)

# test1: sum weight by Level == 100 (sum weight by Level_0 != 100)

tmp1 <- d1 \%>\%
group_by(freq.x, month.yr, Level) \%>\%
mutate(test1 = sum(w2022))

# test2: sum weight by Cat and level
k <- max(d1$qrt.yr) - 1
tmp2 <- d1 \%>\%
filter(freq.x == "quarter",
Level \%in\% c(1, 8),
qrt.yr == k,
wCat == "wSaison") \%>\%
group_by(freq.x, month.yr, Level) \%>\%
mutate(test2 = sum(w2022))

}

}
